+++ app/controllers/concerns/questionnaires_helper.rb
@@ -1,0 +1,20 @@
+module QuestionnairesHelper
+  # Format for Quick-quiz
+  def format_questionnaire(questionnaire)
+    {
+        title: questionnaire.title,
+        questions: questionnaire.questions.each_with_index.map do |question, i|
+          {
+              number: i + 1,
+              prompt: question.text,
+              answers: question.answers.map(&:text),
+              correct: {
+                  index: question.answers.each_with_index.find { |answer, _|
+                    answer.is_correct
+                  }[1]
+              }
+          }
+        end
+    }
+  end
+end
