+++ app/controllers/gifts_controller.rb
@@ -1,4 +1,6 @@
 class GiftsController < ApiController
+  include QuestionnairesHelper
+
   def index
     token = params[:token]
     if token.nil?
@@ -18,25 +20,5 @@ class GiftsController < ApiController
         charityProject: gift.donation.charity_project,
         questionnaire: format_questionnaire(gift.donation.charity_project.questionnaire)
     }, status: :ok
-  end
-
-  private
-  # Format for Quick-quiz
-  def format_questionnaire(questionnaire)
-    {
-        title: questionnaire.title,
-        questions: questionnaire.questions.each_with_index.map do |question, i|
-          {
-              number: i + 1,
-              prompt: question.text,
-              answers: question.answers.map(&:text),
-              correct: {
-                  index: question.answers.each_with_index.find { |answer, _|
-                    answer.is_correct
-                  }[1]
-              }
-          }
-        end
-    }
   end
 end
