+++ app/controllers/questionnaires_controller.rb
@@ -1,0 +1,8 @@
+class QuestionnairesController < ApiController
+  include QuestionnairesHelper
+
+  def show
+    formatted_questionnaire = format_questionnaire Questionnaire.find(params[:id])
+    render json: formatted_questionnaire, status: :ok
+  end
+end
