+++ config/routes.rb
@@ -8,5 +8,7 @@ Rails.application.routes.draw do
 
   resources :gifts, only: :index
 
+  resources :questionnaires, only: :show
+
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
